+++ lib/utils/index.js
@@ -2,10 +2,14 @@ const getFilenameFromUrl = require('./getFilenameFromUrl');
 const handleRangeHeaders = require('./handleRangeHeaders');
 const handleRequest = require('./handleRequest');
 const ready = require('./ready');
+const setupOutputFileSystem = require('./setupOutputFileSystem');
+const setupWriteToDisk = require('./setupWriteToDisk');
 
 module.exports = {
   getFilenameFromUrl,
   handleRangeHeaders,
   handleRequest,
   ready,
+  setupOutputFileSystem,
+  setupWriteToDisk,
 };
